Refactor HierTrackSel so its components can be more easily used elsewhere
